third_party/blink/web_tests/fast/canvas-api/OffscreenCanvas-transferable-expected.txt

Test OffscreenCanvas transferable with exception cases.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS offscreenCanvas.width is width
PASS offscreenCanvas.height is height
PASS offscreenCanvas.width is 0
PASS offscreenCanvas.height is 0
PASS newOffscreenCanvas.width is width
PASS newOffscreenCanvas.height is height
PASS successfullyParsed is true

TEST COMPLETE
